Notenarchiv

+++ src/main/java/com/silviomoser/mhz/api/core/CrudApi.java
@@ -30,6 +30,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 public abstract class CrudApi<T extends AbstractEntity> {
 
 
+
+
     @Autowired
     protected AbstractCrudService<T> service;
 
@@ -40,7 +42,7 @@ public abstract class CrudApi<T extends AbstractEntity> {
             @ApiResponse(code = 200, message = "Success", response = Article.class),
             @ApiResponse(code = 400, message = "Bad request")
     })
-    public T getById(@PathVariable("id") Integer id) throws CrudServiceException {
+    public T getById(@PathVariable("id") Long id) throws CrudServiceException {
         return service.get(id);
     }
 
@@ -81,7 +83,7 @@ public abstract class CrudApi<T extends AbstractEntity> {
 
 
     @DeleteMapping(value = "/{id}")
-    public void delete(@PathVariable("id") Integer id) throws CrudServiceException {
+    public void delete(@PathVariable("id") Long id) throws CrudServiceException {
         checkNotNull(id);
         service.delete(id);
     }
